Subviews modules refactor
Deleted a package and moved the searchbar template to the other package to reduce code duplication
