+++ app/src/main/java/com/example/exchangeapp/screens/information/subviews/BasicScreenViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.exchangeapp.screens.information.subview2
+package com.example.exchangeapp.screens.information.subviews
 
 import com.example.exchangeapp.screens.ExchangeAppViewModel
 import com.google.firebase.firestore.FirebaseFirestore
@@ -6,8 +6,29 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class SearchBarScreenViewModel @Inject constructor() : ExchangeAppViewModel(){
+class BasicScreenViewModel @Inject constructor() : ExchangeAppViewModel() {
+
     private val db = FirebaseFirestore.getInstance()
+
+    fun getDocument(onDataReceived: (String, String) -> Unit){
+
+        var docRef = db.collection("adapting_tips").document("E7t32f5jyCijoZQXIJsf")
+
+        docRef.get().addOnSuccessListener { document ->
+            if (document != null) {
+                // Accessing the title field
+                var title = document.getString("title")
+                var description = document.getString("description")
+                if (title != null && description != null) {
+                        onDataReceived(title,description)
+                }
+            } else {
+                println("No such document!")
+            }
+        }.addOnFailureListener { exception ->
+            println("Error getting document: $exception")
+        }
+    }
 
     fun getDocumentsData(collection : String,onDataReceived: (List<Map<String, Any>>) -> Unit) {
         val collectionRef = db.collection(collection)
